two fixes
