+++ code/modules/atmospherics/machinery/components/gas_recipe_machines/crystallizer.dm
@@ -292,12 +292,14 @@
 		for(var/gasid in internal.gases)
 			internal_gas_data.Add(list(list(
 			"name"= internal.gases[gasid][GAS_META][META_GAS_NAME],
+			"id" = internal.gases[gasid][GAS_META][META_GAS_ID],
 			"amount" = round(internal.gases[gasid][MOLES], 0.01),
 			)))
 	else
 		for(var/gasid in internal.gases)
 			internal_gas_data.Add(list(list(
 				"name"= internal.gases[gasid][GAS_META][META_GAS_NAME],
+				"id" = internal.gases[gasid][GAS_META][META_GAS_ID],
 				"amount" = 0,
 				)))
 	data["internal_gas_data"] = internal_gas_data
